backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-generativeai==0.8.3
pillow>=11.1.0
python-multipart==0.0.20
python-dotenv==1.0.1
